pyproject.toml
[tool.poetry]
name = "cqtest"
version = "0.1.0"
description = "cqtest"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0.a1"
nb-cli = "^0.1.0"

[tool.poetry.dev-dependencies]
nonebot-test = "^0.1.0"

[nonebot.plugins]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["cqtest/plugins"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
